refactor: contexts 폴더 분리 #429
