+++ frontend/src/hooks/useToast.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import ToastContext from '@/providers/ToastProvider/ToastContext';
+import ToastContext from '@/contexts/ToastContext';
 
 const useToast = () => {
   const toast = useContext(ToastContext);
